Refactored push/clone + factored out authentication related helpers
