--- src/functions/push/helpers/get-protocol.js
@@ -1,10 +1,0 @@
-module.exports = async () => {
-  const git = require('isomorphic-git')
-  const fs = require('fs')
-  const originUrl = await git.getConfig({
-    fs,
-    dir: '.',
-    path: 'remote.origin.url'
-  })
-  return originUrl.substring(0, 5) === 'https' ? 'https' : 'ssh'
-}
