+++ src/functions/push/helpers/git-process/git-process.js
@@ -1,10 +1,9 @@
 module.exports = async (protocol) => {
   const {addAll, params, message} = await require('./info-prompter.js')()
-  let mfa = false
   if (protocol === 'https') {
-    mfa = await require('./https-instructions.js')()
+    await require('../../../../utils/auth/https-instructions.js')()
   }
   await require('./stage.js')(addAll, params)
   await require('./commit.js')(message)
-  await require('./push.js')(protocol, mfa)
+  await require('./push.js')(protocol)
 }
