+++ src/utils/auth/get-protocol.js
@@ -1,0 +1,12 @@
+module.exports = async (url) => {
+  if (!url) {
+    const git = require('isomorphic-git')
+    const fs = require('fs')
+    url = await git.getConfig({
+      fs,
+      dir: '.',
+      path: 'remote.origin.url'
+    })
+  }
+  return url.substring(0, 5) === 'https' ? 'https' : 'ssh'
+}
